third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/css/css-flexbox/abspos/flex-abspos-staticpos-align-self-vertWM-002-expected.txt

This is a testharness.js-based test.
FAIL .container > div 1 assert_equals: 
<div class="container"><div style="align-self: auto" data-offset-x="10" data-offset-y="1"></div></div>
offsetLeft expected 10 but got 1
FAIL .container > div 2 assert_equals: 
<div class="container"><div style="align-self: normal" data-offset-x="10" data-offset-y="1"></div></div>
offsetLeft expected 10 but got 1
FAIL .container > div 3 assert_equals: 
<div class="container"><div style="align-self: stretch" data-offset-x="10" data-offset-y="1"></div></div>
offsetLeft expected 10 but got 1
FAIL .container > div 4 assert_equals: 
<div class="container"><div style="align-self: baseline" data-offset-x="10" data-offset-y="1"></div></div>
offsetLeft expected 10 but got 1
FAIL .container > div 5 assert_equals: 
<div class="container"><div style="align-self: last baseline" data-offset-x="2" data-offset-y="1"></div></div>
offsetLeft expected 2 but got 1
FAIL .container > div 6 assert_equals: 
<div class="container"><div style="align-self: center" data-offset-x="6" data-offset-y="1"></div></div>
offsetLeft expected 6 but got 1
FAIL .container > div 7 assert_equals: 
<div class="container"><div style="align-self: start" data-offset-x="10" data-offset-y="1"></div></div>
offsetLeft expected 10 but got 1
FAIL .container > div 8 assert_equals: 
<div class="container"><div style="align-self: end" data-offset-x="2" data-offset-y="1"></div></div>
offsetLeft expected 2 but got 1
FAIL .container > div 9 assert_equals: 
<div class="container"><div style="align-self: self-start" data-offset-x="2" data-offset-y="1"></div></div>
offsetLeft expected 2 but got 1
FAIL .container > div 10 assert_equals: 
<div class="container"><div style="align-self: self-end" data-offset-x="10" data-offset-y="1"></div></div>
offsetLeft expected 10 but got 1
FAIL .container > div 11 assert_equals: 
<div class="container"><div style="align-self: flex-start" data-offset-x="10" data-offset-y="1"></div></div>
offsetLeft expected 10 but got 1
FAIL .container > div 12 assert_equals: 
<div class="container"><div style="align-self: flex-end" data-offset-x="2" data-offset-y="1"></div></div>
offsetLeft expected 2 but got 1
FAIL .container > div 13 assert_equals: 
<div class="container"><div style="align-self: left" data-offset-x="10" data-offset-y="1"></div></div>
offsetLeft expected 10 but got 1
FAIL .container > div 14 assert_equals: 
<div class="container"><div style="align-self: right" data-offset-x="10" data-offset-y="1"></div></div>
offsetLeft expected 10 but got 1
FAIL .container > div 15 assert_equals: 
<div class="container"><div style="align-self: auto" data-offset-x="-2" data-offset-y="1"></div></div>
offsetLeft expected -2 but got 1
FAIL .container > div 16 assert_equals: 
<div class="container"><div style="align-self: normal" data-offset-x="-2" data-offset-y="1"></div></div>
offsetLeft expected -2 but got 1
FAIL .container > div 17 assert_equals: 
<div class="container"><div style="align-self: stretch" data-offset-x="-2" data-offset-y="1"></div></div>
offsetLeft expected -2 but got 1
FAIL .container > div 18 assert_equals: 
<div class="container"><div style="align-self: baseline" data-offset-x="-2" data-offset-y="1"></div></div>
offsetLeft expected -2 but got 1
FAIL .container > div 19 assert_equals: 
<div class="container"><div style="align-self: last baseline" data-offset-x="2" data-offset-y="1"></div></div>
offsetLeft expected 2 but got 1
FAIL .container > div 20 assert_equals: 
<div class="container"><div style="align-self: center" data-offset-x="0" data-offset-y="1"></div></div>
offsetLeft expected 0 but got 1
FAIL .container > div 21 assert_equals: 
<div class="container"><div style="align-self: start" data-offset-x="-2" data-offset-y="1"></div></div>
offsetLeft expected -2 but got 1
FAIL .container > div 22 assert_equals: 
<div class="container"><div style="align-self: end" data-offset-x="2" data-offset-y="1"></div></div>
offsetLeft expected 2 but got 1
FAIL .container > div 23 assert_equals: 
<div class="container"><div style="align-self: self-start" data-offset-x="2" data-offset-y="1"></div></div>
offsetLeft expected 2 but got 1
FAIL .container > div 24 assert_equals: 
<div class="container"><div style="align-self: self-end" data-offset-x="-2" data-offset-y="1"></div></div>
offsetLeft expected -2 but got 1
FAIL .container > div 25 assert_equals: 
<div class="container"><div style="align-self: flex-start" data-offset-x="-2" data-offset-y="1"></div></div>
offsetLeft expected -2 but got 1
FAIL .container > div 26 assert_equals: 
<div class="container"><div style="align-self: flex-end" data-offset-x="2" data-offset-y="1"></div></div>
offsetLeft expected 2 but got 1
FAIL .container > div 27 assert_equals: 
<div class="container"><div style="align-self: left" data-offset-x="-2" data-offset-y="1"></div></div>
offsetLeft expected -2 but got 1
FAIL .container > div 28 assert_equals: 
<div class="container"><div style="align-self: right" data-offset-x="-2" data-offset-y="1"></div></div>
offsetLeft expected -2 but got 1
Harness: the test ran to completion.
